requirements.txt
asgiref==3.6.0
astroid==2.15.5
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
dill==0.3.6
Django==4.2.1
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
idna==3.4
inflection==0.5.1
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.2
mccabe==0.7.0
packaging==23.1
Pillow==9.5.0
platformdirs==3.5.1
PyJWT==2.7.0
pylint==2.17.4
pylint-django==2.5.3
pylint-plugin-utils==0.8.2
pytz==2023.3
requests==2.31.0
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.11.8
typing_extensions==4.6.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.2
wrapt==1.15.0
